refactor: [#1495] reorganize torrent-repository mod
