--- packages/torrent-repository/src/entry/mod.rs
@@ -1,1 +1,0 @@
-pub mod swarm;
